agregué checkeo de permisos en el controlador de usuarios

+++ admin/src/web/controllers/users.py
@@ -13,9 +13,13 @@ from core.user import create_user, list_roles, list_users, search_users, update_
 bprint = Blueprint("users", __name__, url_prefix="/usuarios")
 
 @bprint.get("/")
-#@login_required
-#@check("user_index")
+
 def index():
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_index"):
+        return abort(403)
     query = request.args.get('query',"")
     role = request.args.get('role', None)
     active = request.args.get('active', None)
@@ -43,6 +47,11 @@ def activar_usuario():
     """
     Funcion que permite habilitar/deshabilitar el usuario a menos que sea System Admin
     """
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_activar_usuario"):
+        return abort(403)
     chosen_id = request.form['id']
     query = request.form['query']
     user = User.query.get(chosen_id)
@@ -64,6 +73,12 @@ def edit_user(user_id):
     """ Función que edita el usuario y agrega los parametros a un diccionario para ahorrar 
         validación en la funcion update_user
     """
+
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_edit_user"):
+        return abort(403)
     user = User.query.get_or_404(user_id)  
     roles = Role.query.all()  
 
@@ -85,6 +100,13 @@ def edit_user(user_id):
 
 @bprint.post("/delete_user")
 def delete_user_controller():
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_delete_user_controller"):
+        return abort(403)
+    
+
     user_id = request.form.get("user_id")  
     try:
         delete_user(user_id) 
@@ -96,6 +118,13 @@ def delete_user_controller():
 
 @bprint.route("/register_user", methods=["GET", "POST"])
 def register_user():
+
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_register_user"):
+        return abort(403)
+
     roles = list_roles()
     roles = list_roles()
